refactor: extract Homebrew platform infrastructure
